requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
slack-bolt>=1.18.0
slack-sdk>=3.26.0
pyyaml>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
